# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Botica SaaS API
  version: 1.0.0
  description: API multiempresa para operacion de boticas.
paths:
...
  /api/v1/cash-sessions/{id}/counts/:
    get:
      operationId: api_v1_cash_sessions_counts_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: Un Cadena UUID que identifique este cash session.
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashSession'
          description: ''
    post:
      operationId: api_v1_cash_sessions_counts_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: Un Cadena UUID que identifique este cash session.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashSession'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CashSession'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CashSession'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashSession'
          description: ''
...
components:
  schemas:
...
    CashMovement:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        performed_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        movement_type:
          allOf:
          - $ref: '#/components/schemas/CashMovementMovementTypeEnum'
          readOnly: true
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        reference_type:
          type: string
          readOnly: true
        reference_id:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        reason:
          type: string
          readOnly: true
        session:
          type: string
          format: uuid
          readOnly: true
        performed_by:
          type: integer
          readOnly: true
      required:
      - amount
      - created_at
      - id
      - movement_type
      - performed_by
      - performed_by_name
      - reason
      - reference_id
      - reference_type
      - session
      - updated_at
    CashMovementMovementTypeEnum:
      enum:
      - OPENING
      - CASH_SALE
      - MANUAL_IN
      - WITHDRAWAL
      - REFUND
      - RECEIVABLE_COLLECTION
      - PAYABLE_PAYMENT
      - EXPENSE
      - ADDITIONAL_INCOME
      - CONTRIBUTION
      type: string
      description: |-
        * `OPENING` - Apertura
        * `CASH_SALE` - Venta en efectivo
        * `MANUAL_IN` - Ingreso manual
        * `WITHDRAWAL` - Retiro
        * `REFUND` - Devolucion
        * `RECEIVABLE_COLLECTION` - Cobro de cuenta
        * `PAYABLE_PAYMENT` - Pago de cuenta
        * `EXPENSE` - Gasto
        * `ADDITIONAL_INCOME` - Ingreso adicional
        * `CONTRIBUTION` - Aporte de caja
...
    CashSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        register_name:
          type: string
          readOnly: true
        opened_by_name:
          type: string
          readOnly: true
        movements:
          type: array
          items:
            $ref: '#/components/schemas/CashMovement'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/CashSessionStatusEnum'
          readOnly: true
        opening_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        expected_cash:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        counted_cash:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
        difference:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
        opened_at:
          type: string
          format: date-time
          readOnly: true
        closed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        notes:
          type: string
          readOnly: true
        register:
          type: string
          format: uuid
          readOnly: true
        opened_by:
          type: integer
          readOnly: true
        closed_by:
          type: integer
          readOnly: true
          nullable: true
      required:
      - closed_at
      - closed_by
      - counted_cash
      - created_at
      - difference
      - expected_cash
      - id
      - movements
      - notes
      - opened_at
      - opened_by
      - opened_by_name
      - opening_amount
      - register
      - register_name
      - status
      - updated_at
    CashSessionStatusEnum:
      enum:
      - OPEN
      - CLOSING
      - CLOSED
      - CLOSED_WITH_DIFFERENCE
      type: string
      description: |-
        * `OPEN` - Abierta
        * `CLOSING` - En cierre
        * `CLOSED` - Cerrada
        * `CLOSED_WITH_DIFFERENCE` - Cerrada con diferencia
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT